requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Migrate
Flask-Login
python-dotenv
psycopg2-binary 
pymysql